Add files via upload
After defining a function, we use the function by typing out the name of the function, optionally followed by listing the arguments to give (or “pass”) to the function to use as values for any parameters it defined. We also need to use any value returned by the function, by assigning it a variable, or perhaps using that value again as argument for another function.